Split conversions into three files
