+++ crates/store/re_protos/src/protobuf_conversions/log_types.rs
@@ -1,44 +1,3 @@
-use arrow::{datatypes::Schema as ArrowSchema, error::ArrowError};
-
-impl TryFrom<crate::common::v0::Schema> for ArrowSchema {
-    type Error = ArrowError;
-
-    fn try_from(value: crate::common::v0::Schema) -> Result<Self, Self::Error> {
-        Ok(Self::clone(
-            re_sorbet::schema_from_ipc(&value.arrow_schema)?.as_ref(),
-        ))
-    }
-}
-
-impl TryFrom<ArrowSchema> for crate::common::v0::Schema {
-    type Error = ArrowError;
-
-    fn try_from(value: ArrowSchema) -> Result<Self, Self::Error> {
-        Ok(Self {
-            arrow_schema: re_sorbet::ipc_from_schema(&value)?,
-        })
-    }
-}
-
-// -------------------------------------------------------------------------------------------
-
-impl From<crate::common::v0::Tuid> for re_tuid::Tuid {
-    fn from(value: crate::common::v0::Tuid) -> Self {
-        Self::from_nanos_and_inc(value.time_ns, value.inc)
-    }
-}
-
-impl From<re_tuid::Tuid> for crate::common::v0::Tuid {
-    fn from(value: re_tuid::Tuid) -> Self {
-        Self {
-            time_ns: value.nanoseconds_since_epoch(),
-            inc: value.inc(),
-        }
-    }
-}
-
-// -------------------------------------------------------------------------------------------
-
 use crate::TypeConversionError;
 use crate::{invalid_field, missing_field};
 use std::sync::Arc;
@@ -542,13 +501,6 @@ impl TryFrom<crate::log_msg::v0::BlueprintActivationCommand>
 
 #[cfg(test)]
 mod tests {
-    #[test]
-    fn test_tuid_conversion() {
-        let tuid = re_tuid::Tuid::new();
-        let proto_tuid: crate::common::v0::Tuid = tuid.into();
-        let tuid2: re_tuid::Tuid = proto_tuid.into();
-        assert_eq!(tuid, tuid2);
-    }
 
     #[test]
     fn entity_path_conversion() {
